requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
distro==1.9.0
docstring-parser==0.16
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.3
openai==1.61.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
referencing==0.36.2
rpds-py==0.22.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.12.2